+++ tests/OperatingSystemTest.php
@@ -14,7 +14,7 @@ use Innmind\OperatingSystem\{
 };
 use Innmind\Server\Status\Server as ServerStatus;
 use Innmind\Server\Control\Server as ServerControl;
-use Innmind\TimeContinuum\Clock;
+use Innmind\Time\Clock;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class OperatingSystemTest extends TestCase