+++ tests/RemoteTest.php
@@ -11,12 +11,7 @@ use Innmind\OperatingSystem\{
 };
 use Innmind\Server\Control\{
     Server,
-    Servers,
-    Server\Processes,
-    Server\Volumes,
     Server\Command,
-    Server\Process\Pid,
-    Server\Signal,
 };
 use Innmind\Url\{
     Url,
@@ -29,7 +24,6 @@ use Innmind\IO\Sockets\{
 use Innmind\IP\IPv4;
 use Innmind\HttpTransport\Transport as HttpTransport;
 use Innmind\Immutable\Attempt;
-use Formal\AccessLayer\Connection;
 use Innmind\BlackBox\{
     PHPUnit\BlackBox,
     PHPUnit\Framework\TestCase,
@@ -51,8 +45,8 @@ class RemoteTest extends TestCase
 
         $remoteServer = $remote->ssh(Url::of('ssh://user@my-vps:42/'));
 
-        $this->assertInstanceOf(Servers\Remote::class, $remoteServer);
-        $remoteServer
+        $this->assertInstanceOf(Server::class, $remoteServer);
+        $_ = $remoteServer
             ->processes()
             ->execute(Command::foreground('ls'))
             ->unwrap();
@@ -67,8 +61,8 @@ class RemoteTest extends TestCase
 
         $remoteServer = $remote->ssh(Url::of('ssh://user@my-vps:42/'));
 
-        $this->assertInstanceOf(Servers\Logger::class, $remoteServer);
-        $remoteServer
+        $this->assertInstanceOf(Server::class, $remoteServer);
+        $_ = $remoteServer
             ->processes()
             ->execute(Command::foreground('ls'))
             ->unwrap();
@@ -83,8 +77,11 @@ class RemoteTest extends TestCase
 
         $remoteServer = $remote->ssh(Url::of('ssh://user@my-vps/'));
 
-        $this->assertInstanceOf(Servers\Remote::class, $remoteServer);
-        $remoteServer->processes()->execute(Command::foreground('ls'));
+        $this->assertInstanceOf(Server::class, $remoteServer);
+        $_ = $remoteServer
+            ->processes()
+            ->execute(Command::foreground('ls'))
+            ->unwrap();
     }
 
     public function testSocket()
@@ -107,8 +104,8 @@ class RemoteTest extends TestCase
         );
 
         $this->assertInstanceOf(Client::class, $socket);
-        $server->close();
-        $socket->close();
+        $_ = $server->close()->unwrap();
+        $_ = $socket->close()->unwrap();
     }
 
     public function testHttp(): BlackBox\Proof
@@ -140,66 +137,23 @@ class RemoteTest extends TestCase
             ->prove(function($server, $os) {
                 $sql = $os->remote()->sql($server);
 
-                $this->assertInstanceOf(Connection::class, $sql);
+                $this->assertInstanceOf(Attempt::class, $sql);
             });
     }
 
     private function server(string ...$commands): Server
     {
-        return new class($this->processes(), $this, $commands) implements Server {
-            private $inner;
+        $processes = Factory::build()->control()->processes();
 
-            public function __construct(
-                private $processes,
-                private $test,
-                private $commands,
-            ) {
-            }
+        return Server::via(function($command) use (&$commands, $processes) {
+            $expected = \array_shift($commands);
+            $this->assertNotNull($expected);
+            $this->assertSame(
+                $expected,
+                $command->toString(),
+            );
 
-            public function processes(): Processes
-            {
-                return $this->inner ??= new class($this->processes, $this->test, $this->commands) implements Processes {
-                    public function __construct(
-                        private $processes,
-                        private $test,
-                        private $commands,
-                    ) {
-                    }
-
-                    public function execute(Command $command): Attempt
-                    {
-                        $expected = \array_shift($this->commands);
-                        $this->test->assertNotNull($expected);
-                        $this->test->assertSame(
-                            $expected,
-                            $command->toString(),
-                        );
-
-                        return $this->processes->execute(Command::foreground('echo'));
-                    }
-
-                    public function kill(Pid $pid, Signal $signal): Attempt
-                    {
-                    }
-                };
-            }
-
-            public function volumes(): Volumes
-            {
-            }
-
-            public function reboot(): Attempt
-            {
-            }
-
-            public function shutdown(): Attempt
-            {
-            }
-        };
-    }
-
-    private function processes(): Processes
-    {
-        return Factory::build()->control()->processes();
+            return $processes->execute(Command::foreground('echo'));
+        });
     }
 }